Tweaking carrying capacity calculations
Estimates of the global population of earth 10,000 years ago vary wildly, but you could pick out a mean of 10 million people if you wanted to. Avar has 61% of the surface area of earth, so we might want to shoot for 6 million people. A carrying capacity factor of 1/6 gets us there.