Change elf generation from 10 to 5
Elves could have kids at 100, but it's like humans having kids at 20; it's the lower bound of acceptable. We count generations as 25 in the ancient world or 30 today, because a lot of people have kids later. So we should bump elf generations up to 200 years, or 5 times in a millennium.