OH GODS WHAT HAVE I DONE
This commit is an abomination in the eyes of gods and men. Across the endless possibilities of existence, there must be a place where justice dwells, and in that place I shall suffer eternal torment for what I have done here. This is a single commit in which I rewrite a huge portion of the codebase. May whatever gods may hear the pitiful cries of a sinner as wretched as I have mercy on my miserable soul.